requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Settings
pydantic-settings==2.6.1
pydantic[email]>=2.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector>=0.2.5,<0.3.0

# MinIO
minio==7.2.11

# Auth (Keycloak JWT)
python-jose[cryptography]==3.3.0
httpx==0.27.2

# LangChain / LangGraph / RAG
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.11
langgraph==0.2.56
langchain-postgres==0.0.12

# File parsing (Python 3.10+ required for markitdown)
markitdown[all]==0.1.1
python-multipart==0.0.18

# Test
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Lint / Format
ruff==0.8.3

# Security scan
pip-audit==2.7.3
